+++ src/pybindings/pybindings_learning/pybindings_independences.cpp
@@ -59,7 +59,7 @@ void pybindings_independence_tests(py::module& root) {
         .def("column_names", &LinearCorrelation::column_names);
 
     py::class_<KMutualInformation, IndependenceTest, std::shared_ptr<KMutualInformation>>(independence_tests, "KMutualInformation")
-        .def(py::init<DataFrame, int, long unsigned int, int, int>(),
+        .def(py::init<DataFrame, int, unsigned int, int, int>(),
             py::arg("df"), py::arg("k"), py::arg("seed"), py::arg("shuffle_neighbors") = 5, py::arg("samples") = 1000)
         .def(py::init<DataFrame, int,  int, int>(),
             py::arg("df"), py::arg("k"), py::arg("shuffle_neighbors") = 5, py::arg("samples") = 1000)
